dom/worklet/tests/test_dump.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Worklet - Console</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="common.js"></script>
</head>
<body>

<script type="application/javascript">
// This function is called into an iframe.
function runTestInIframe() {
  var audioContext = new AudioContext();
  audioContext.audioWorklet.addModule("worklet_dump.js")
  .then(() => {
    ok(true, "All good!");
    SimpleTest.finish();
  });
}
</script>
</body>
</html>
